ISPN-10673 Remove Cross-Site Replication from Local caches
