+++ core/src/test/resources/configs/xsite/xsite-inline-test.xml
@@ -66,12 +66,12 @@
 
    <cache-container default-cache="default">
       <transport cluster="infinispan-cluster" lock-timeout="50000" stack="xsite" node-name="Jalapeno" machine="m1" rack="r1" site="LON"/>
-      <local-cache name="default">
+      <distributed-cache name="default">
          <backups>
             <backup site="NYC" strategy="SYNC" failure-policy="IGNORE" timeout="12003"/>
             <backup site="SFO" strategy="ASYNC"/>
          </backups>
-      </local-cache>
+      </distributed-cache>
    </cache-container>
 
 </infinispan>
