+++ core/src/test/resources/configs/xsite/xsite-offline-test.xml
@@ -3,29 +3,29 @@
 
    <cache-container default-cache="default">
       <transport cluster="infinispan-cluster" lock-timeout="50000" stack="udp" node-name="Jalapeno" machine="m1" rack="r1" site="LON"/>
-      <local-cache-configuration name="base">
+      <distributed-cache-configuration name="base">
          <backups>
             <backup site="NYC" failure-policy="IGNORE" strategy="SYNC" timeout="12003" enabled="true">
                <take-offline after-failures="123" min-wait="5673"/>
             </backup>
          </backups>
          <backup-for remote-cache="someCache" remote-site="SFO"/>
-      </local-cache-configuration>
-      <local-cache name="default" configuration="base"/>
-      <local-cache name="inheritor" configuration="base"/>
-      <local-cache name="noTakeOffline" configuration="base">
+      </distributed-cache-configuration>
+      <distributed-cache name="default" configuration="base"/>
+      <distributed-cache name="inheritor" configuration="base"/>
+      <distributed-cache name="noTakeOffline" configuration="base">
          <backups>
             <backup site="NYC" failure-policy="WARN" strategy="SYNC" timeout="12003" enabled="true" />
          </backups>
          <backup-for/>
-      </local-cache>
-      <local-cache name="takeOfflineDifferentConfig" configuration="base">
+      </distributed-cache>
+      <distributed-cache name="takeOfflineDifferentConfig" configuration="base">
          <backups>
             <backup site="NYC" failure-policy="IGNORE" strategy="SYNC" timeout="12003" enabled="true">
                <take-offline after-failures="321" min-wait="3765"/>
             </backup>
          </backups>
-      </local-cache>
+      </distributed-cache>
    </cache-container>
 
 </infinispan>
