+++ core/src/test/resources/configs/xsite/xsite-test2.xml
@@ -3,25 +3,25 @@
 
    <cache-container default-cache="default">
       <transport cluster="infinispan-cluster" lock-timeout="50000" stack="udp" node-name="Jalapeno" machine="m1" rack="r1" site="LON"/>
-      <local-cache-configuration name="base">
+      <distributed-cache-configuration name="base">
          <backups>
             <backup site="NYC" failure-policy="IGNORE" strategy="SYNC" timeout="12003" enabled="true"/>
             <backup site="SFO" strategy="ASYNC" enabled="true"/>
             <backup site="LON" strategy="SYNC" enabled="false" />
          </backups>
          <backup-for remote-cache="someCache" remote-site="SFO"/>
-      </local-cache-configuration>
-      <local-cache name="default" configuration="base"/>
-      <local-cache name="inheritor" configuration="base"/>
-      <local-cache name="noBackupFor" configuration="base">
+      </distributed-cache-configuration>
+      <distributed-cache name="default" configuration="base"/>
+      <distributed-cache name="inheritor" configuration="base"/>
+      <distributed-cache name="noBackupFor" configuration="base">
          <backups>
             <backup site="NYC" strategy="SYNC" failure-policy="WARN" timeout="12003"/>
          </backups>
          <backup-for/>
-      </local-cache>
-      <local-cache name="noBackupFor2" configuration="base">
+      </distributed-cache>
+      <distributed-cache name="noBackupFor2" configuration="base">
          <backups/>
-      </local-cache>
+      </distributed-cache>
    </cache-container>
 
 </infinispan>
